+++ src/ui/AppBar.js
@@ -1,0 +1,18 @@
+import React from 'react';
+import { Box } from 'grommet';
+
+const AppBar = (props) => (
+    <Box
+        tag='header'
+        direction='row'
+        align='center'
+        justify='between'
+        background='brand'
+        pad={{ left: 'medium', right: 'small', vertical: 'small' }}
+        elevation='medium'
+        style={{ zIndex: '1' }}
+        {...props}
+    />
+);
+
+export default AppBar;
